requirements.txt
click
face_recognition
face_recognition_models
gradio==5.44.1
numpy
onnxruntime_gpu
opencv_python
opencv_python_headless
Pillow
requests
torch
tqdm
